Move models to seperate folder.
