+++ gridt/src/app/core/models/signal.model.ts
