+++ gridt/src/app/core/swap.service.ts
@@ -2,8 +2,8 @@
  * A service for storing the swap events, created for solving issue #54.
  */
 import { Injectable } from '@angular/core';
-import { Movement } from './movement.model';
-import { User } from './user.model';
+import { Movement } from './models/movement.model';
+import { User } from './models/user.model';
 
 export interface SwapEvent {
   date: Date;
@@ -14,7 +14,7 @@ export interface SwapEvent {
 @Injectable({
   providedIn: "root"
 })
-export class SwapService {  
+export class SwapService {
   public swapEvents: SwapEvent[] = [];
 
   addSwapEvent (movement: Movement, user: User) {
@@ -34,8 +34,8 @@ export class SwapService {
     if (!events.length) {
       return null;
     }
-    
-    const lastSwapEvent = events.reduce( 
+
+    const lastSwapEvent = events.reduce(
       (prev, current) => (prev.date > current.date) ? prev : current
     );
