+++ gridt/src/app/movements/movements-detail/movements-detail.page.ts
@@ -6,7 +6,7 @@ import { Observable } from 'rxjs';
 import { map, flatMap, take } from 'rxjs/operators';
 
 import { ApiService } from '../../core/api.service';
-import { Movement } from '../../core/movement.model';
+import { Movement } from '../../core/models/movement.model';
 
 @Component({
   selector: 'app-movements-detail',
@@ -27,10 +27,10 @@ export class MovementsDetailPage implements OnInit {
   ngOnInit() {
     const id = this.route.snapshot.paramMap.get('movementId');
     this.movement$ = this.api.allMovements$.pipe(
-      map(movements => movements.filter( movement => movement.name === id )[0]) 
+      map(movements => movements.filter( movement => movement.name === id )[0])
     );
   }
-  
+
   onSubscribe(): void {
     this.alertCtrl.create({
       header: 'Are you sure?',
@@ -83,7 +83,7 @@ export class MovementsDetailPage implements OnInit {
     el.present();
 
     this.movement$.pipe(
-      take(1), 
+      take(1),
       flatMap( (movement) => this.api.subscribeToMovement$(movement.name)),
       flatMap( () => this.movement$), // We need the movement name again to reload the movement.
       take(1),
